Add executions relation to Task entity
Introduced a new OneToMany relation for 'executions' in the Task entity to enable tracking of task executions. Updated the query builder in the Task repository to join with the executions, as well as with assignee and workingOn for enriched data retrieval. This enhancement aims to improve task tracking and reporting functionality.